pyproject.toml
[project]
name = "access_mopper"

authors = [
    {name = "Romain Beucher", email = "romain.beucher@anu.edu.au"},
    {name = "Paola Petrelli", email = "paola.petrelli@utas.edu.au"},
    {name = "Sam Green", email = "sam.green@unsw.edu.au"},
]
description = "ACCESS Model Output Post-Processor, maps raw model output to CMIP-style defined variables and produce post-processed output using CMOR3"
readme = "README.md"
requires-python = ">=3.11"
keywords = ["ACCESS", "post-processing"]
license = {text = "Apache-2.0"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "cmor",
    "numpy",
    "pandas",
    "xarray",
    "dask",
    "cftime",
    "pyyaml",
]
dynamic = ["version"]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-cov",
    "ruff"
]

[build-system]
build-backend = "setuptools.build_meta"
requires = [
  "setuptools>=64.0.0",
  "versioneer[toml]",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
mopdata = ["*.json", "*.yaml", "*.csv",
           "cmor_tables/*.json"]

[tool.versioneer]
VCS = "git"
style = "pep440"
versionfile_source = "src/access_mopper/_version.py"
versionfile_build = "access_mopper/_version.py"
tag_prefix = ""
parentdir_prefix = "access_mopper-"

[tool.ruff]
exclude = ["versioneer.py"]
line-length = 88  # Matches Black's default
target-version = "py311"  # Adjust to match your minimum Python version

[tool.ruff.lint]
select = ["E", "F", "W", "I"]  # E: pycodestyle, F: Pyflakes, W: warnings, I: isort
ignore = ["E203", "E501"]  # Ignore line-length issues (handled by formatter)

[tool.ruff.format]
quote-style = "double"  # Use double quotes
